meme_gen/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
djangorestframework-simplejwt==5.3.1
pyjwt==2.9.0
Pillow==10.1.0
requests==2.32.3
urllib3==2.2.3